+++ modelopt/torch/export/tensorrt_llm_utils.py
@@ -50,6 +50,7 @@ MODEL_NAME_TO_HF_ARCH_MAP = {
     "enc": "EncoderModel",
     "dec": "DecoderModel",
     "mllama": "MLLaMAModel",
+    "whisper_encoder": "WhisperEncoder",
 }
 
 
@@ -121,7 +122,9 @@ def convert_to_tensorrt_llm_config(
     if is_enc_dec:
         # For encoder
         if model_config.enc_dec == "enc":
-            config_architecture = MODEL_NAME_TO_HF_ARCH_MAP["enc"]
+            config_architecture = MODEL_NAME_TO_HF_ARCH_MAP[
+                f"{decoder_type}_encoder" if decoder_type in ["whisper"] else "enc"
+            ]
         # For decoder
         else:
             config_architecture = MODEL_NAME_TO_HF_ARCH_MAP["dec"]
@@ -185,7 +188,7 @@ def convert_to_tensorrt_llm_config(
             from tensorrt_llm.models import MODEL_MAP
             from tensorrt_llm.models.modeling_utils import Mapping, QuantConfig
 
-            model_cls = MODEL_MAP[model_config.architecture]
+            model_cls = MODEL_MAP[config_architecture]
             config_cls = getattr(model_cls, "config_class", None)
 
             if config_cls is not None:
@@ -249,6 +252,12 @@ def convert_to_tensorrt_llm_config(
         config["emb_scale_by_sqrt_dim"] = model_config.layers[0].emb_scale_by_sqrt_dim
         config["layer_types"] = model_config.layers[0].layer_types
     elif is_enc_dec:
+        if decoder_type in ["whisper"]:
+            config["n_mels"] = hf_config.num_mel_bins
+            model_is_multilingual = hf_config.vocab_size >= 51865
+            num_languages = hf_config.vocab_size - 51765 - int(model_is_multilingual)
+            config["num_languages"] = num_languages
+
         # T5 models use relative position embedding
         # Bart models use learned_absolute position embedding
         config["position_embedding_type"] = (
@@ -312,7 +321,6 @@ def convert_to_tensorrt_llm_config(
             config["eos_token_id"] = model_config.eos_token_id
             config["bos_token_id"] = model_config.bos_token_id
             config["pad_token_id"] = model_config.pad_token_id
-
     elif decoder_type == "dbrx":
         config["clip_qkv"] = first_attention_decoder_config.clip_qkv
 
@@ -372,7 +380,7 @@ def convert_to_tensorrt_llm_config(
 def prepare_enc_dec_export_dir(tensorrt_llm_config: dict[str, Any], export_root: Path):
     """Prepare the export directory for encoder-decoder model."""
     # For encoder
-    if tensorrt_llm_config["architecture"] == "EncoderModel":
+    if tensorrt_llm_config["architecture"] in ["EncoderModel", "WhisperEncoder"]:
         export_dir = export_root.joinpath("encoder")
     # For decoder
     else: